refactor UnitIO, adding UEnv and more
Units can now end the enclosing chain via a standard UEnv object. U and
UChain are changed accordingly. A UChain checks internally which units
may end it, and it sets the correct doneActions before playing back or
when releasing. The duration of a UEnv can be inf. U:release will
release the envelope of the unit and use the fadeOut time by default (or
a user given time, analogous to Synth:release).