adding shouldPlayOn for units and UChains
Any unit in a UChain can now block playback for a specific target or
server. It does so based on the shouldNotPlayOnFunc in Udef. If the func
{ |unit, target| … } returns false the playback is blocked, unless there
are any other units in the chain that return true. In the last case only
the unit doesn't play, and the rest of the chain does. The function is
called both at prepare and makeSynth. If the function is nil (default)
or returns nil that means that the other units in the chain may decide
wether to play or not. If all units return nil the chain is played.