adding CPULoadBalancer
a class tracking groups of servers based on estimated cpu, which gives
the opportunity to fetch the server with the lowest cpu count within a
group. The method to use normally is Server.nextLowestCPU( addAmt, group
) or the dedicated server instance method which knows it's own group
(may return another server from that group). This returns a new server
and immediately adds an estimated cpu count to it, so that the next time
another one is chosen. Not entirely sure yet where to put this in the
UChain system.