adding WFSPanSynth, updating WFSArrayPan
WFSArrayPan for points now fully works. WFSPanSynth (or actually
WFSArrayPanSynth) makes the synthdefs for the array panniers. It takes
half a minute on my machine to generate all defs. We should not do this
automatically, I think only on request. The defs don't need to be
updated when the speakerconf is changed, so they probably almost never
will have to be generated again (unless their code has changed or the
arrays are more than 64 speakers wide - the default maximum). Perhaps we
should ship the def files with the distro (so they are always loaded at
startup). Luckily once they are generated, they are only 2.7MB of
.scsyndef files (104 defs in total), which are loaded to the server
almost instantly. The defs are also too big to send directly to the
synths, so they have to be loaded on the servers.  Note:
WFSArrayPanPlane is not working correctly yet, although the synthdefs
are already generated. It does take the apx amount of cpu of the final
version already. Check the load panner synthdefs.scd file to see some
examples.