SyncCenter: added flag checkForServerQuit, bypassing quit detection
default is false. I find this too dangerous to be true by default. The
server may very well be out of sight for 5s or more when loading a large
soundfile, or during other heavy network traffic, causing a \didQuit and
subsequent loss of sync data. This will immediately disable playback,
which I find a much bigger problem than possibly not being sure if a
server is still running. The latter is only interesting when developing,
hence the flag. What we could also do is react only if an actual /quit
message is received from the server, then we can be sure it has stopped.
In normal situations the server is never quit.  I also wish a similar
flag would be on Server itself; it should keep notification up until we
say it's over; we are depending on n_free messages for the freeing of
buffers. Chances are that the server is still notified while the lang
doesn't know.