enabling overlapping playback for single Units and chains
AbstractRichBuffer.free buffer now only frees the oldest buffer on a
server, and when creating a new one doesn't do a resetBuffers first.  It
now allows multiple instances of a buffer on a single server. This
enables units and chains to overlap. While this is not intended
behavior, it can easily and often happen with long fadeout times. This
commit will make sure that an ending fadeout from a previous playback of
a unit doesn't interfere with the buffer of a newly played one. The only
moment where this goes wrong is when the playback instance of a unit
that started later, ends earlier. In such cases still all buffers will
be freed eventually, but in the wrong order. In other words we assume
that a chain that was started earlier also ends earlier. It's not
perfect, but at least it does allow overlapping, where before all
buffers were freed at the moment one playback instance ended.