introduce rounding for SyncCenter
It seems required to round off all the sample scheduled messages to the
blocksize of the servers. The disadvantage is that all start times are
rounded to the blocksize (default: 128), but the advantage is that n_set
messages always arrive in the same block as where the synth was started,
therefore preventing things to go out of sync for example when changing
the soundFile rate when a chain is already playing. If people want to
get rid of the rounding they can simply do: SyncCenter.round = 0.   The
rounding is done on the master sample count, which acts as a reference
here.