creating mass editor for units
MassEditU creates a unit that can set all mass-editable args for a
number of units at once. All units need to be from the same udder (the
class will warn and fail if not). In the gui the MassEditU behaves as
if it is a normal unit. Check the mass edit example.scd file in Test
Code.