Refactoring of UScore playing methods
Introduce a "pos" variable that keeps track of the current position. 

The play related methods are now:

//stop just the spawning and releasing of events, i.e., the taks
stopScore

//stop synths
stopChains

//stop everything
stop
	
//pause only if it's unpaused and playing
pause

//resume only if it's paused and playing	
resume

This should make it more clear what the methods actually do, and also enforce that you can't unpause a score that is not paused, which would just play it twice at the same time, which I don't think is a desired result.

prStart now has a task to update that just sends the current position to interested objects. This is used to update the transport bar and time counter (and could be used for other things). In a non gui situations this task can be turned off.

isPlaying, isPaused variables allow querying the current state of the score. 

When copying scores using deepCopy, we need to watch out for the play related variables, they should all be reinitialized.