adding wfsStaticIndex and wfsDynamicIndex Udefs
These replace the 'index' mode of WFSSynth in v1. Lets you play a sound
on one single speaker in the system. The static version plays only on
the server where that drives the speaker, but the index itself still
can be changed during playback (i.e. it is not really static). It will
mute when out of range.