enabling startPos for U, UChain and soundfile playback
example:

u = UChain( "sounds/a11wlk01.wav", \output );
u.prepareAndStart( s, 1 ); // start one second into the file

the startPos can be passed on to the unit's synthdef via the
\u_startPos control. BufSndFilePlayer already uses this internally.
This means that still the whole buffer is loaded, but only the part
after the startPos is played *).

Also any \u_dur args will be altered, as well as \u_fadeIn. Both the
prepare and start functions have a startPos argument. The prepare
function passes the startpos on to all .prepare calls. This is used by
DiskSndFile to set the read position to the correct value.

*) in the future we could make startPos to be used intelligently on
BufSndFile: when loop == false we actually don't need to load the whole
soundfile, when loop == true we do..

todo: startPos for WFSPathBufferPlayer