Copy/Paste and MenuBar
 - Copy/Paste support between scores and inside a score added-
 - WFSMenubar for commands relating to the current score.
 - UScoreView - added many methods related with manipulating currently selected events.
 - MouseManager - if no events selected now returns nil, in order to use !? on the result

Note started using the !? operator which applied the function to the value if it's not nil and does nothing if it's nil. This is very useful to have chains of actions on values that might be nil. For the moment there's a huge amount of "function not inlined" warnings, still checking how to solve that. The warning can always be turned off.