Refactoring
It's now possible to open sub-scores in sub windows.
Made sure that controllers are properly removed when windows are closed.
Some refactoring of "changed" message names.
UScore - added notification of when the number of events changes. This is useful to just recreate the eventViews if new events are added or removed.