refactor WFSPath2
WFSPath2 now knows it's own file name. It also knows if it was saved or
not. asCompileString returns different things according to if fileName
is known and if the saved state is equal to the current. Still it is
not entirely deepCopy proof yet; copying a WFSPathBuffer will also copy
the path, and subsequently loose it's filepath. A filePath normally can
only be linked to one WFSPath2.