UScore - saving/dirty status
UScore now keeps track if it's contents changed since last time it changed by keeping a copy of itself at that time as a string. UScore:save will not save if nothing changed.
When closing the score editor GUI it will check if the score is dirty, and only ask for save if it is.