prevent crash in UChainGUI
apparently when a widget inside the new scrollview is focused at the
moment of removal, sc crashes. I'm focusing the start button now before
removing the composite. It is a bit hacky but does seem to do the
trick. The bug typically showed when selecting a soundfile with a
different number of channels in a soundfile playback unit. But also
sometimes when replacing a Udef while one of the parameter sliders was
focused.