adding WFSPathGeneratorView to editor and renaming / moving WFSPath related classes adding WFSPathGeneratorView to editor and renaming / moving WFSPath related classes
renaming:
- WFSPathEditor is now WFSPathGUI
- SimpleEdit (a single editing function) -> SimpleTransformer
- WFSPathEdit is now WFSPathTransformer
- WFSPathEditView -> WFSPathTransformerView

when clicking the "generate" button in WFSPathGUI a
WFSPathGeneratorView window is called, linked to the current
WFSPathGUI. For now it holds a fixed set of generators, which will be
changeable in the future

bug: when moving the "blend" slider for a single generator it seems to
replace the points in it's own copy. I'm not sure why, have to
investigate. It should keep it's own internal copy of the original in
tact so that when moving the slider back to 0 it gets the original path
back