correction for plane waves on smaller systems
this is a temporary fix; the attenuation needs to be global (the same
for all speakers in a system), but that requires some changes in the
architecture. This fix will work best for systems with equal amounts of
speakers on each array, and less well systems with big differences in
the amount of speakers per array.