WFSLib.startup can now be used instead of WFS.startup
- users can now put a file called "preferences.scd" either in the
appsupport dir or next to the app itself. The file is loaded before
startup, and may create a WFSSpeakerConf, a WFSOptions and possibly
other settings that are system-specific.
- the parsing of old WFSCollider_configuration.txt files is untested
- the presets for bea7 are untested, and probably wrong
- at current, the servers of the game of life system will not produce a
sound anymore after startup. This feature might be re-introduced,
although the synthdef writing now happens before startup, so basically
the system is ready as soon as all servers are started.