WFS panner defs from zip file in the Resources folder at first run.
in the current macOS applications from unknown developers are started
from a scrambled write-only path, which makes it impossible to write
synthdefs inside an application bundle. Therefore the synthdefs folder
for WFSCollider standalone is now moved to userAppSupportDir and the
synthdefs are copied there at first run from a zip file inside the
resource dir.