fix node allocator issue modern SC
since "users" were introduced for remote servers a problem with high node ID's surfaced. This is now solved by only allowing nodeID's that are compatible with 32bits float values. Current max-users is set to 2 to allow stil 8 million node ID's to be allocated. Each step increase of the maxUsers will halve that value. Probably up to 4 or 5 is ok, but above that may become problematic for long-lasting playback.